Install devel packages only if needed

+++ rh-build.sh
@@ -10,10 +10,7 @@ then
     pkgcmd="yum"
     sed -i '/^tsflags=/d' /etc/yum.conf
     yum install -y epel-release
-    yum install -y @buildsys-build
-    yum install -y yum-utils
     yum install -y yum-plugin-copr
-    yum install -y git
     for copr in $COPR
     do
         yum copr enable -y $copr epel-7
@@ -24,11 +21,6 @@ then
     builddep="dnf builddep"
     sed -i '/^tsflags=/d' /etc/dnf/dnf.conf
     dnf install -y epel-release
-    dnf install -y 'dnf-command(config-manager)'
-    dnf config-manager --set-enabled PowerTools
-    dnf groupinstall -y "Development Tools"
-    dnf install -y 'dnf-command(builddep)'
-    dnf install -y git
     for copr in $COPR
     do
         dnf copr enable -y $copr
@@ -37,9 +29,6 @@ elif [[ "$distro" =~ fedora ]]
 then
     pkgcmd="dnf"
     sed -i '/^tsflags=/d' /etc/dnf/dnf.conf
-    dnf install -y --allowerasing @buildsys-build
-    dnf install -y 'dnf-command(builddep)'
-    dnf install -y git
     for copr in $COPR
     do
         dnf copr enable -y $copr
@@ -49,22 +38,27 @@ else
     exit 1
 fi
 
-mkdir src
-pushd src
+if [[ ${#GITREF[@]} -gt 0 ]]
+then
+    $pkgcmd install -y git
 
-for repo in wreport dballe meteo-vm2 arkimet libsim
-do
-    ref=${GITREF[$repo]:-}
-    [[ -z "$ref" ]] && continue
-    git clone http://github.com/arpa-simc/$repo
-    pushd $repo
-    git checkout $ref
-    bash .travis-build.sh $distro
-    find ~/rpmbuild/RPMS -name "*.rpm" -print0 | xargs -0 $pkgcmd install -y
-    popd # $repo
-done
+    mkdir src
+    pushd src
 
-popd # src
-rm -rf src
+    for repo in wreport dballe meteo-vm2 arkimet libsim
+    do
+        ref=${GITREF[$repo]:-}
+        [[ -z "$ref" ]] && continue
+        git clone http://github.com/arpa-simc/$repo
+        pushd $repo
+        git checkout $ref
+        bash .travis-build.sh $distro
+        find ~/rpmbuild/RPMS -name "*.rpm" -print0 | xargs -0 $pkgcmd install -y
+        popd # $repo
+    done
+
+    popd # src
+    rm -rf src
+fi
 
 $pkgcmd install -y ${PACKAGES}
